requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
google-genai==1.13.0
playwright==1.52.0
lxml==5.4.0
crawl4ai==0.6.2
python-dateutil==2.9.0.post0
pymongo==4.12.1
python-dotenv==1.1.0
pydantic-setting==2.9.1
pydantic==2.11.4
pydantic_core==2.33.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.4.0